iso_deposit_extractors subdirectory in EgammaIsolationAlgos
