+++ RecoEgamma/EgammaIsolationAlgos/plugins/EgammaHcalPFClusterIsolationProducer.cc
@@ -1,23 +1,57 @@
-#include "RecoEgamma/EgammaIsolationAlgos/plugins/EgammaHcalPFClusterIsolationProducer.h"
-#include "HLTrigger/HLTcore/interface/defaultModuleLabel.h"
-
-#include "FWCore/Utilities/interface/TypeID.h"
-
-#include "FWCore/Framework/interface/EventSetup.h"
-#include "DataFormats/Common/interface/Handle.h"
-#include "FWCore/Framework/interface/ESHandle.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
+//*****************************************************************************
+// File:      EgammaHcalPFClusterIsolationProducer.cc
+// ----------------------------------------------------------------------------
+// OrigAuth:  Matteo Sani
+// Institute: UCSD
+//*****************************************************************************
 
 #include "DataFormats/Candidate/interface/CandAssociation.h"
+#include "DataFormats/Common/interface/Handle.h"
 #include "DataFormats/EgammaCandidates/interface/GsfElectron.h"
 #include "DataFormats/EgammaCandidates/interface/Photon.h"
-
+#include "DataFormats/ParticleFlowReco/interface/PFCluster.h"
+#include "DataFormats/ParticleFlowReco/interface/PFClusterFwd.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
+#include "FWCore/Utilities/interface/TypeID.h"
+#include "HLTrigger/HLTcore/interface/defaultModuleLabel.h"
 #include "RecoEgamma/EgammaIsolationAlgos/interface/HcalPFClusterIsolation.h"
 
-#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
-#include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
-
 #include <typeinfo>
+
+template <typename T1>
+class EgammaHcalPFClusterIsolationProducer : public edm::stream::EDProducer<> {
+public:
+  typedef std::vector<T1> T1Collection;
+  typedef edm::Ref<T1Collection> T1Ref;
+  explicit EgammaHcalPFClusterIsolationProducer(const edm::ParameterSet&);
+  ~EgammaHcalPFClusterIsolationProducer() override;
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+  void produce(edm::Event&, const edm::EventSetup&) override;
+
+private:
+  const edm::EDGetTokenT<T1Collection> emObjectProducer_;
+  const edm::EDGetTokenT<reco::PFClusterCollection> pfClusterProducerHCAL_;
+  const edm::EDGetTokenT<reco::PFClusterCollection> pfClusterProducerHFEM_;
+  const edm::EDGetTokenT<reco::PFClusterCollection> pfClusterProducerHFHAD_;
+
+  const bool useHF_;
+  const double drMax_;
+  const double drVetoBarrel_;
+  const double drVetoEndcap_;
+  const double etaStripBarrel_;
+  const double etaStripEndcap_;
+  const double energyBarrel_;
+  const double energyEndcap_;
+  const double useEt_;
+};
 
 template <typename T1>
 EgammaHcalPFClusterIsolationProducer<T1>::EgammaHcalPFClusterIsolationProducer(const edm::ParameterSet& config)
